eoscompanion: Break service up into different files
