+++ eoscompanion/util.py
@@ -1,0 +1,29 @@
+# /eoscompanion/util.py
+#
+# Copyright (C) 2017, 2018 Endless Mobile, Inc.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+# All rights reserved.
+'''Logging and utility functions for eoscompanion.'''
+
+import sys
+
+from gi.repository import GLib
+
+
+def log(msg, file=sys.stdout):
+    '''Log the message to the console if not running quietly.'''
+    if not GLib.getenv('EOS_COMPANION_APP_SERVICE_QUIET'):
+        print(msg, file=file)
